requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────────
Django==6.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1

# ── Database ──────────────────────────────────────────────────────────────────
psycopg2-binary==2.9.9

# ── Media & Storage ───────────────────────────────────────────────────────────
Pillow>=10.4.0
django-storages==1.14.2
boto3==1.34.69

# ── Security ──────────────────────────────────────────────────────────────────
# Field-level encryption for bank accounts, national IDs
cryptography==42.0.5
# Input sanitisation
bleach==6.1.0
# File MIME type validation (requires: brew install libmagic)
python-magic==0.4.27
# 2FA
pyotp==2.9.0
# Google OAuth verification
google-auth==2.29.0
# Secret detection
detect-secrets==1.4.0

# ── Redis, Cache, Channels ────────────────────────────────────────────────────
redis==5.0.3
django-redis==5.4.0
channels==4.0.0
channels-redis==4.2.0
daphne==4.1.2

# ── Celery ────────────────────────────────────────────────────────────────────
celery==5.3.6
django-celery-beat==2.6.0

# ── Structured Logging ────────────────────────────────────────────────────────
python-json-logger==2.0.7

# ── Push Notifications ────────────────────────────────────────────────────────
firebase-admin==6.4.0

# ── Error Tracking ────────────────────────────────────────────────────────────
sentry-sdk==1.45.0

# ── API Docs ──────────────────────────────────────────────────────────────────
drf-spectacular==0.27.2

# ── PDF Generation ────────────────────────────────────────────────────────────
weasyprint==61.2

# ── Environment ───────────────────────────────────────────────────────────────
python-decouple==3.8

# ── Retry / Circuit Breaker ───────────────────────────────────────────────────
tenacity==8.3.0

# ── HTTP client ───────────────────────────────────────────────────────────────
requests==2.31.0
